move Blob/File and optimize the buffer module
